Implemented the code for an univariete linear regression

+++ Linear-Regression/CPP/mean-squared-error.cpp
@@ -8,10 +8,10 @@
     for a given values of the modal parameters. The formula is,
     J(p1, p2,..., pn) = (1/2m)*(sum((prediction[i] - true_value[i])**2))
 */
-double Mean_Squared_Error::calculate_error(std::vector<double> y_hat, std::vector<double> y) {
-    double m = y.size();
-    double error = 0, square_diff = 0;
-
+float Mean_Squared_Error::calculate_error(std::vector<float> y_hat, std::vector<float> y) {
+    float m = y.size();
+    float error = 0, square_diff = 0;
+    
     for (int i = 0; i < m; i++) {
         square_diff += std::pow(y_hat[i] - y[i], 2);
